+++ Tests/SentryTests/SentryOptionsTest.m
@@ -503,6 +503,8 @@ typedef SentryLog *_Nullable (^SentryBeforeSendLogCallback)(SentryLog *_Nonnull 
     XCTAssertEqual(nil, options.beforeBreadcrumb);
 }
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
 - (void)testOnCrashedLastRun
 {
     __block BOOL onCrashedLastRunCalled = NO;
@@ -529,6 +531,30 @@ typedef SentryLog *_Nullable (^SentryBeforeSendLogCallback)(SentryLog *_Nonnull 
     SentryOptions *options = [self getValidOptions:@{ @"onCrashedLastRun" : @"fault" }];
 
     XCTAssertNil(options.onCrashedLastRun);
+}
+#pragma clang diagnostic pop
+
+- (void)testOnLastRunStatus
+{
+    void (^callback)(NSInteger, SentryEvent *_Nullable)
+        = ^(__unused NSInteger status, __unused SentryEvent *_Nullable event) { };
+    SentryOptions *options = [self getValidOptions:@{ @"onLastRunStatusDetermined" : callback }];
+
+    XCTAssertNotNil(options.onLastRunStatusDetermined);
+}
+
+- (void)testDefaultOnLastRunStatus
+{
+    SentryOptions *options = [self getValidOptions:@{}];
+
+    XCTAssertNil(options.onLastRunStatusDetermined);
+}
+
+- (void)testGarbageOnLastRunStatus_ReturnsNil
+{
+    SentryOptions *options = [self getValidOptions:@{ @"onLastRunStatusDetermined" : @"fault" }];
+
+    XCTAssertNil(options.onLastRunStatusDetermined);
 }
 
 - (void)testSampleRateWithDict
@@ -704,7 +730,11 @@ typedef SentryLog *_Nullable (^SentryBeforeSendLogCallback)(SentryLog *_Nonnull 
     XCTAssertTrue([[self getDefaultCacheDirectoryPath] isEqualToString:options.cacheDirectoryPath]);
     XCTAssertNil(options.beforeSend);
     XCTAssertNil(options.beforeBreadcrumb);
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     XCTAssertNil(options.onCrashedLastRun);
+#pragma clang diagnostic pop
+    XCTAssertNil(options.onLastRunStatusDetermined);
     XCTAssertEqual(1.0, options.sampleRate.floatValue);
     XCTAssertEqual(YES, options.enableAutoSessionTracking);
     XCTAssertEqual(YES, options.enableWatchdogTerminationTracking);